feat-update role based access

+++ lib/features/projects/domain/use_cases/delete_project_use_case.dart
@@ -1,5 +1,7 @@
 import 'package:fpdart/fpdart.dart';
+import 'package:grubpac/core/error/common_failures.dart';
 import 'package:grubpac/core/error/failure.dart';
+import 'package:grubpac/core/shared/enums.dart';
 import 'package:grubpac/core/usecase/use_case.dart';
 import 'package:grubpac/features/projects/domain/params/project_params.dart';
 import 'package:grubpac/features/projects/domain/repo/i_projects_repo.dart';
@@ -13,6 +15,15 @@ class DeleteProjectUseCase implements UseCase<void, DeleteProjectParams> {
   Future<Either<Failure, void>> call({
     required DeleteProjectParams parameters,
   }) {
+    if (parameters.session.role != UserRole.orgAdmin) {
+      return Future.value(
+        left(
+          ForbiddenFailure(
+            message: 'Only organization admins can delete projects.',
+          ),
+        ),
+      );
+    }
     return repository.deleteProject(
       projectId: parameters.projectId,
       session: parameters.session,
